+++ Sources/SwifterSwift/UIKit/UIImageExtensions.swift
@@ -141,7 +141,7 @@ public extension UIImage {
                                      width: destRect.width.rounded(),
                                      height: destRect.height.rounded())
 
-        UIGraphicsBeginImageContext(roundedDestRect.size)
+        UIGraphicsBeginImageContextWithOptions(roundedDestRect.size, false, scale)
         guard let contextRef = UIGraphicsGetCurrentContext() else { return nil }
 
         contextRef.translateBy(x: roundedDestRect.width / 2, y: roundedDestRect.height / 2)
@@ -171,7 +171,7 @@ public extension UIImage {
                                      width: destRect.width.rounded(),
                                      height: destRect.height.rounded())
 
-        UIGraphicsBeginImageContext(roundedDestRect.size)
+        UIGraphicsBeginImageContextWithOptions(roundedDestRect.size, false, scale)
         guard let contextRef = UIGraphicsGetCurrentContext() else { return nil }
 
         contextRef.translateBy(x: roundedDestRect.width / 2, y: roundedDestRect.height / 2)